Update the specification to use HTTP HEAD instead of HTTP GET
Chromium initiates the download of a payment method manifest through
HTTP HEAD and looks for the Link HTTP header:
https://crrev.com/c/4954394

This patch updates the specification to replace the HTTP GET with HTTP
HEAD in the first request for the payment method manifest download. This
aligns the specification more closely to the implementation in Chromium.

(Chromium and other user agents based on it are the only implementers of
this specification at this time.)

Fixed: https://crbug.com/384016068